create log folder when not exist

+++ src/lib/logger.ts
@@ -1,14 +1,20 @@
 import * as path from "path";
-import { deleteOldFiles } from "./utils";
+import { deleteOldFiles, checkDirectory } from "./utils";
 import pino from "pino";
 
 export const defaultLogFolder = path.join("logs");
 
 let logFileStream: pino.Logger | undefined;
 
-export function createLogger() {
+export async function createLogger() {
   const fileName = `${new Date().getTime()}.log`;
   const filePath = path.join(defaultLogFolder, fileName);
+
+  /** make sure logs folder exist */
+  const logFolderExist = await checkDirectory(defaultLogFolder);
+  if (!logFolderExist) {
+    throw new Error("Logs Folder not Exist");
+  }
 
   deleteOldFiles({ options: { keepMetaFiles: true, keepAliveTime: 60 } });
